# packages/mtmai/mtmai-0.3.1218-py3-none-any.whl/mtmai/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "mtmai",
    "description": "mtmai description(group)",
    "version": "0.3.1218"
  },
  "paths": {
...
    "/api/v1/chat/config": {
      "post": {
        "tags": [
          "openai"
        ],
        "summary": "Chat Config",
        "operationId": "openai-chat_config",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatConfigRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssisantConfig"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/v1/chat/completions": {
      "post": {
        "tags": [
          "openai"
        ],
        "summary": "Chat Completions",
        "description": "\u4ee5\u517c\u5bb9 openai completions \u534f\u8bae\u7684\u65b9\u5f0f \u53cd\u4ee3\u5176\u4ed6 ai \u63d0\u4f9b\u5546\u7684 completions api",
        "operationId": "openai-chat_completions",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AssisantConfig": {
        "properties": {
          "chatProfile": {
            "type": "string",
            "title": "Chatprofile"
          },
          "logo": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Logo"
          },
          "welcome": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AssisantWelcome"
              },
              {
                "type": "null"
              }
            ]
          },
          "starts": {
            "items": {
              "$ref": "#/components/schemas/AssisantStart"
            },
            "type": "array",
            "title": "Starts"
          },
          "workbench": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AssisantWorkbenchConfig"
              },
              {
                "type": "null"
              }
            ]
          },
          "context": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AssisantContext"
              },
              {
                "type": "null"
              }
            ]
          },
          "menus": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/AssisantMenus"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Menus"
          },
          "siderView": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SiderViewConfig"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "chatProfile",
          "starts"
        ],
        "title": "AssisantConfig"
      },
      "AssisantContext": {
        "properties": {
          "siteId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Siteid"
          },
          "userId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Userid"
          },
          "threadId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Threadid"
          },
          "params": {
            "anyOf": [
              {
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Params"
          }
        },
        "type": "object",
        "title": "AssisantContext"
      },
      "AssisantMenus": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "label": {
            "type": "string",
            "title": "Label"
          },
          "icon": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Icon"
          },
          "viewName": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Viewname"
          },
          "viewProps": {
            "anyOf": [
              {
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Viewprops"
          },
          "children": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/AssisantMenus"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Children"
          }
        },
        "type": "object",
        "required": [
          "id",
          "label"
        ],
        "title": "AssisantMenus",
        "description": "\u5e94\u7528\u4e3b\u83dc\u5355"
      },
      "AssisantStart": {
        "properties": {
          "title": {
            "type": "string",
            "title": "Title"
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message"
          },
          "icon": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Icon"
          }
        },
        "type": "object",
        "required": [
          "title"
        ],
        "title": "AssisantStart"
      },
      "AssisantWelcome": {
        "properties": {
          "title": {
            "type": "string",
            "title": "Title"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          }
        },
        "type": "object",
        "required": [
          "title"
        ],
        "title": "AssisantWelcome"
      },
      "AssisantWorkBrench": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "label": {
            "type": "string",
            "title": "Label"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "icon": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Icon"
          }
        },
        "type": "object",
        "required": [
          "id",
          "label",
          "description"
        ],
        "title": "AssisantWorkBrench"
      },
      "AssisantWorkbenchConfig": {
        "properties": {
          "workbenchs": {
            "items": {
              "$ref": "#/components/schemas/AssisantWorkBrench"
            },
            "type": "array",
            "title": "Workbenchs"
          },
          "workbench_default": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Workbench Default"
          }
        },
        "type": "object",
        "required": [
          "workbenchs"
        ],
        "title": "AssisantWorkbenchConfig"
      },
...
      "ChatConfigRequest": {
        "properties": {
          "chatProfile": {
            "type": "string",
            "title": "Chatprofile"
          }
        },
        "type": "object",
        "required": [
          "chatProfile"
        ],
        "title": "ChatConfigRequest"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "SiderViewConfig": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "label": {
            "type": "string",
            "title": "Label"
          },
          "icon": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Icon"
          },
          "viewName": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Viewname"
          },
          "viewProps": {
            "anyOf": [
              {
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Viewprops"
          }
        },
        "type": "object",
        "required": [
          "id",
          "label"
        ],
        "title": "SiderViewConfig",
        "description": "\u4fa7\u8fb9\u680f\u89c6\u56fe\u914d\u7f6e"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/v1/login/access-token"
          }
        }
      }
    }
  },
...
}